requirements.txt
Django>=4.1,<4.2
psycopg2>=2.8
black>=22.6.0
djangorestframework>=3.13,<3.14
isort
pre-commit
